Create update database structure with QGIS 3.16

+++ gobs/processing/algorithms/database/create.py
@@ -1,7 +1,6 @@
-__copyright__ = "Copyright 2020, 3Liz"
+__copyright__ = "Copyright 2022, 3Liz"
 __license__ = "GPL version 3"
 __email__ = "info@3liz.org"
-__revision__ = "$Format:%H$"
 
 import os
 
@@ -11,30 +10,25 @@ from qgis.core import (
     QgsProcessingOutputNumber,
     QgsProcessingOutputString,
     QgsProcessingParameterBoolean,
-    QgsProcessingParameterString,
+    QgsProcessingParameterProviderConnection,
     QgsProject,
+    QgsProviderConnectionException,
+    QgsProviderRegistry,
 )
 
-from gobs.qgis_plugin_tools.tools.algorithm_processing import (
-    BaseProcessingAlgorithm,
-)
-from gobs.qgis_plugin_tools.tools.i18n import tr
-from gobs.qgis_plugin_tools.tools.resources import (
+from gobs.plugin_tools import (
+    available_migrations,
     plugin_path,
     plugin_test_data_path,
 )
+from gobs.processing.algorithms.database.base import BaseDatabaseAlgorithm
+from gobs.qgis_plugin_tools.tools.i18n import tr
 from gobs.qgis_plugin_tools.tools.version import version
-
-from ...qgis_plugin_tools.tools.database import (
-    available_migrations,
-    fetch_data_from_sql_query,
-)
-from .tools import get_postgis_connection_list
 
 SCHEMA = "gobs"
 
 
-class CreateDatabaseStructure(BaseProcessingAlgorithm):
+class CreateDatabaseStructure(BaseDatabaseAlgorithm):
     """
     Create gobs structure in Database
     """
@@ -52,12 +46,6 @@ class CreateDatabaseStructure(BaseProcessingAlgorithm):
     def displayName(self):
         return tr('Create database structure')
 
-    def group(self):
-        return tr('Structure')
-
-    def groupId(self):
-        return 'gobs_structure'
-
     def shortHelpString(self):
         short_help = tr(
             'Install the G-Obs database structure with tables and function on the chosen database.'
@@ -74,21 +62,18 @@ class CreateDatabaseStructure(BaseProcessingAlgorithm):
         return short_help
 
     def initAlgorithm(self, config):
-        # INPUTS
+
         project = QgsProject.instance()
         connection_name = QgsExpressionContextUtils.projectScope(project).variable('gobs_connection_name')
-        db_param = QgsProcessingParameterString(
+        param = QgsProcessingParameterProviderConnection(
             self.CONNECTION_NAME,
-            tr('PostgreSQL connection to G-Obs database'),
+            tr("Connection to the PostgreSQL database"),
+            "postgres",
             defaultValue=connection_name,
-            optional=False
+            optional=False,
         )
-        db_param.setMetadata({
-            'widget_wrapper': {
-                'class': 'processing.gui.wrappers_postgis.ConnectionWidgetWrapper'
-            }
-        })
-        self.addParameter(db_param)
+        param.setHelp(tr("The database where the schema '{}' will be installed.").format(SCHEMA))
+        self.addParameter(param)
 
         self.addParameter(
             QgsProcessingParameterBoolean(
@@ -120,57 +105,47 @@ class CreateDatabaseStructure(BaseProcessingAlgorithm):
         )
 
     def checkParameterValues(self, parameters, context):
-        # Check that the connection name has been configured
-        connection_name = parameters[self.CONNECTION_NAME]
-        if not connection_name:
-            return False, tr('You must use the "Configure G-obs plugin" alg to set the database connection name')
+        metadata = QgsProviderRegistry.instance().providerMetadata("postgres")
+        connection_name = self.parameterAsConnectionName(
+            parameters, self.CONNECTION_NAME, context
+        )
+        connection = metadata.findConnection(connection_name)
+        override = self.parameterAsBoolean(parameters, self.OVERRIDE, context)
 
-        # Check that it corresponds to an existing connection
-        if connection_name not in get_postgis_connection_list():
-            return False, tr('The configured connection name does not exists in QGIS')
-
-        # Check database content
-        ok, msg = self.checkSchema(parameters, context)
-        if not ok:
+        if 'gobs' in connection.schemas() and not override:
+            msg = tr("Schema gobs already exists in database ! If you REALLY want to drop and recreate it (and loose all data), check the *Overwrite* checkbox")
             return False, msg
+
         return super(CreateDatabaseStructure, self).checkParameterValues(parameters, context)
 
-    def checkSchema(self, parameters, context):
-        sql = '''
-            SELECT schema_name
-            FROM information_schema.schemata
-            WHERE schema_name = 'gobs';
-        '''
-        connection_name = parameters[self.CONNECTION_NAME]
-        [header, data, rowCount, ok, error_message] = fetch_data_from_sql_query(
-            connection_name,
-            sql
-        )
-        if not ok:
-            return ok, error_message
-        override = parameters[self.OVERRIDE]
-        msg = tr('Schema gobs does not exists. Continue...')
-        for a in data:
-            schema = a[0]
-            if schema == 'gobs' and not override:
-                ok = False
-                msg = tr("Schema gobs already exists in database ! If you REALLY want to drop and recreate it (and loose all data), check the *Overwrite* checkbox")
-        return ok, msg
-
     def processAlgorithm(self, parameters, context, feedback):
-        connection_name = parameters[self.CONNECTION_NAME]
+
+        metadata = QgsProviderRegistry.instance().providerMetadata("postgres")
+        connection_name = self.parameterAsConnectionName(
+            parameters, self.CONNECTION_NAME, context
+        )
+
+        # noinspection PyTypeChecker
+        connection = metadata.findConnection(connection_name)
+        if not connection:
+            raise QgsProcessingException(
+                f"La connexion {connection_name} n'existe pas."
+            )
 
         # Drop schema if needed
         override = self.parameterAsBool(parameters, self.OVERRIDE, context)
         if override:
             feedback.pushInfo(tr("Trying to drop schema {}…").format(SCHEMA))
             sql = "DROP SCHEMA IF EXISTS {} CASCADE;".format(SCHEMA)
-
-            _, _, _, ok, error_message = fetch_data_from_sql_query(connection_name, sql)
-            if ok:
-                feedback.pushInfo(tr("Schema {} has been dropped.").format(SCHEMA))
-            else:
-                raise QgsProcessingException(error_message)
+            metadata = QgsProviderRegistry.instance().providerMetadata("postgres")
+            connection_name = self.parameterAsConnectionName(
+                parameters, self.CONNECTION_NAME, context
+            )
+            try:
+                connection.executeSql(sql)
+            except QgsProviderConnectionException as e:
+                raise QgsProcessingException(str(e))
+            feedback.pushInfo("  Success !")
 
         # Create full structure
         sql_files = [
@@ -194,7 +169,7 @@ class CreateDatabaseStructure(BaseProcessingAlgorithm):
         plugin_version = version()
         dev_version = False
         run_migration = os.environ.get(
-            "TEST_DATABASE_INSTALL_{}".format(SCHEMA.capitalize())
+            "TEST_DATABASE_INSTALL_{}".format(SCHEMA.upper())
         )
         if plugin_version in ["master", "dev"] and not run_migration:
             feedback.reportError(
@@ -220,13 +195,12 @@ class CreateDatabaseStructure(BaseProcessingAlgorithm):
                     feedback.pushInfo("  Skipped (empty file)")
                     continue
 
-                _, _, _, ok, error_message = fetch_data_from_sql_query(
-                    connection_name, sql
-                )
-                if ok:
-                    feedback.pushInfo("  Success !")
-                else:
-                    raise QgsProcessingException(error_message)
+                try:
+                    connection.executeSql(sql)
+                except QgsProviderConnectionException as e:
+                    raise QgsProcessingException(str(e))
+
+                feedback.pushInfo("  Success !")
 
         # Add version
         if run_migration or not dev_version:
@@ -248,7 +222,11 @@ class CreateDatabaseStructure(BaseProcessingAlgorithm):
             SCHEMA, metadata_version
         )
 
-        fetch_data_from_sql_query(connection_name, sql)
+        try:
+            connection.executeSql(sql)
+        except QgsProviderConnectionException as e:
+            raise QgsProcessingException(str(e))
+
         feedback.pushInfo(
             "Database version '{}'.".format(metadata_version)
         )
